upd: updated algorithm and added optuna optimization

+++ extract_grid_good.py
@@ -3,6 +3,8 @@ import numpy as np
 import random
 from itertools import combinations
 import argparse
+import os, time, json
+import optuna
 
 def resize_image_to_fit(image, max_width=1280, max_height=720):
     """
@@ -17,60 +19,165 @@ def resize_image_to_fit(image, max_width=1280, max_height=720):
     scale = min(max_width / width, max_height / height)
     return cv2.resize(image, (int(width * scale), int(height * scale)))
 
-def preprocess_image(image_path):
+def preprocess_image2(image_path, params, use_debug):
+    img = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
+    
+    # CLAHE
+    if params["clahe"]["to_use"]:
+        clahe = cv2.createCLAHE(clipLimit=params["clipLimit"], tileGridSize=(params["clahe"]["tileGridSize"], 
+                                                                             params["clahe"]["tileGridSize"]))
+        img = clahe.apply(img)
+        if use_debug:
+            cv2.imshow("clahe", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
+    # Гауссово размытие
+    if params["gauss"]["to_use"]:
+        img = cv2.GaussianBlur(img, (params["gauss"]["kernel"], params["gauss"]["kernel"]), 0)
+        if use_debug:
+            cv2.imshow("GaussianBlur", resize_image_to_fit(img))
+            cv2.waitKey(0)
+        
+    if params["adaptive_thresh"]["to_use"]:
+        # Адаптивная бинаризация
+        img = cv2.adaptiveThreshold(
+            img, 255, 
+            cv2.ADAPTIVE_THRESH_GAUSSIAN_C, 
+            cv2.THRESH_BINARY_INV, params["adaptive_thresh"]["blockSize"], 
+                                   params["adaptive_thresh"]["const"]
+        )
+        if use_debug:
+            cv2.imshow("adaptiveThreshold", resize_image_to_fit(img))
+            cv2.waitKey(0)
+
+    # img = cv2.bitwise_not(img) 
+    # if use_debug:
+    #     cv2.imshow("bitwise_not", resize_image_to_fit(img))
+    #     cv2.waitKey(0)
+
+    contours, _ = cv2.findContours(img, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
+    for cnt in contours:
+        if cv2.contourArea(cnt) < 2000:
+            cv2.drawContours(img, [cnt], 0, 0, -1)
+    if use_debug:
+        cv2.imshow("findContours", resize_image_to_fit(img))
+        cv2.waitKey(0)
+
+    if params["dilate"]["to_use"]:
+        # Закрытие: соединяем близко расположенные компоненты в единую область
+        kernel_dilate = cv2.getStructuringElement(cv2.MORPH_RECT, (params["dilate"]["kernel"], 
+                                                                   params["dilate"]["kernel"]))
+        img = cv2.dilate(img, kernel_dilate, iterations=params["dilate"]["iterations"])
+        if use_debug:
+            cv2.imshow("dilate", resize_image_to_fit(img))
+            cv2.waitKey(0)
+
+    if params["erode"]["to_use"]:
+        kernel_erode = cv2.getStructuringElement(cv2.MORPH_RECT, (params["erode"]["kernel"], 
+                                                                  params["erode"]["kernel"]))
+        img = cv2.erode(img, kernel_erode, iterations=params["erode"]["iterations"])
+        if use_debug:
+            cv2.imshow("erode", resize_image_to_fit(img))
+            cv2.waitKey(0)
+    
+    # Детекция линий с оптимизированными параметрами
+    lines = cv2.HoughLinesP(
+        img, 
+        rho=1, 
+        theta=np.pi/180, 
+        threshold=params["hough"]["threshold"], 
+        minLineLength=params["hough"]["minLineLength"], 
+        maxLineGap=params["hough"]["maxLineGap"]
+    )
+    
+    return img, lines
+
+def preprocess_image(image_path, params, use_debug):
+    start_time = time.time()
+    # print("preprocess_image started")
     # Загрузка изображения
     img = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
     if img is None:
         print("Не удалось загрузить изображение.")
         exit()
 
-    clahe = cv2.createCLAHE(clipLimit=3.0, tileGridSize=(3, 3))
-    cl1 = clahe.apply(img)
-    cv2.imshow("clahe", resize_image_to_fit(cl1))
-    cv2.waitKey(0)
+    if params["clahe"]["to_use"]:
+        clahe = cv2.createCLAHE(clipLimit=params["clahe"]["clipLimit"], 
+                                tileGridSize=(params["clahe"]["tileGridSize"], 
+                                              params["clahe"]["tileGridSize"]))
+        img = clahe.apply(img)
+        if use_debug:
+            cv2.imshow("clahe", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (3, 3))
+    if params["close1"]["to_use"]:
+        kernel_close1 = cv2.getStructuringElement(cv2.MORPH_RECT, (params["close1"]["kernel"], 
+                                                                   params["close1"]["kernel"]))
+        img = cv2.morphologyEx(img, cv2.MORPH_CLOSE, kernel_close1, iterations=params["close1"]["iterations"])
+        if use_debug:
+            cv2.imshow("closed1", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    opened = cv2.morphologyEx(cl1, cv2.MORPH_CLOSE, kernel, iterations=1)
-    cv2.imshow("opened", resize_image_to_fit(opened))
-    cv2.waitKey(0)
+    if params["gauss"]["to_use"]:
+        # Применяем гауссово размытие для уменьшения шума
+        img = cv2.GaussianBlur(img, (params["gauss"]["kernel"], 
+                                          params["gauss"]["kernel"]), 0)
+        if use_debug:
+            cv2.imshow("blur", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    # Применяем гауссово размытие для уменьшения шума
-    blur = cv2.GaussianBlur(opened, (3, 3), 0)
-    cv2.imshow("blur", resize_image_to_fit(blur))
-    cv2.waitKey(0)
+    if params["canny"]["to_use"]:
+        # Детектор границ Canny
+        img = cv2.Canny(img, params["canny"]["left"], 
+                                params["canny"]["right"], 
+                                apertureSize=3)
+        if use_debug:
+            cv2.imshow("edges", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    # Детектор границ Canny
-    edges = cv2.Canny(blur, 40, 70, apertureSize=3)
-    cv2.imshow("edges", resize_image_to_fit(edges))
-    cv2.waitKey(0)
+    if params["erode1"]["to_use"]:
+        kernel_erode1 = cv2.getStructuringElement(cv2.MORPH_RECT, (params["erode1"]["kernel"], 
+                                                                   params["erode1"]["kernel"]))
+        img = cv2.erode(img, kernel_erode1, iterations=params["erode1"]["iterations"])
+        if use_debug:
+            cv2.imshow("erode1", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    # kernel_erode = cv2.getStructuringElement(cv2.MORPH_RECT, (5, 5))
-    # edges = cv2.erode(edges, kernel_erode, iterations=1)
-    # cv2.imshow("erode1", resize_image_to_fit(edges))
-    # cv2.waitKey(0)
+    if params["dilate"]["to_use"]:
+        # Дилатация: расширяем линии, чтобы группы линий сливались
+        kernel_dilate = cv2.getStructuringElement(cv2.MORPH_RECT, (params["dilate"]["kernel"], 
+                                                                   params["dilate"]["kernel"]))
+        img = cv2.dilate(img, kernel_dilate, iterations=params["dilate"]["iterations"])
+        if use_debug:
+            cv2.imshow("dilated", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    # Дилатация: расширяем линии, чтобы группы линий сливались
+    if params["close2"]["to_use"]:
+        # Закрытие: соединяем близко расположенные компоненты в единую область
+        kernel_close2 = cv2.getStructuringElement(cv2.MORPH_RECT, (params["close2"]["kernel"], 
+                                                                   params["close2"]["kernel"]))
+        img = cv2.morphologyEx(img, cv2.MORPH_OPEN, kernel_close2, iterations=params["close2"]["iterations"])
+        if use_debug:
+            cv2.imshow("closed", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
-    dilated = cv2.dilate(edges, kernel, iterations=1)
-    cv2.imshow("dilated", resize_image_to_fit(dilated))
-    cv2.waitKey(0)
-
-    # Закрытие: соединяем близко расположенные компоненты в единую область
-    closed = cv2.morphologyEx(dilated, cv2.MORPH_OPEN, kernel, iterations=1)
-    cv2.imshow("closed", resize_image_to_fit(closed))
-    cv2.waitKey(0)
-
-    # Эрозия: уменьшаем размер линий
-    kernel_erode = cv2.getStructuringElement(cv2.MORPH_RECT, (2, 2))
-    erode = cv2.erode(closed, kernel_erode, iterations=2)
-    cv2.imshow("erode2", resize_image_to_fit(erode))
-    cv2.waitKey(0)
+    if params["erode2"]["to_use"]:
+        # Эрозия: уменьшаем размер линий
+        kernel_erode2 = cv2.getStructuringElement(cv2.MORPH_RECT, (params["erode2"]["kernel"], 
+                                                                   params["erode2"]["kernel"]))
+        img = cv2.erode(img, kernel_erode2, iterations=params["erode2"]["iterations"])
+        if use_debug:
+            cv2.imshow("erode2", resize_image_to_fit(img))
+            cv2.waitKey(0)
 
     # Поиск линий с использованием преобразования Хафа
-    lines = cv2.HoughLinesP(erode, 1, np.pi / 180, threshold=500, minLineLength=300, maxLineGap=50)
-    return erode, lines
+    lines = cv2.HoughLinesP(img, rho=1, 
+                                    theta=np.pi / 180, 
+                                    threshold=params["hough"]["threshold"], 
+                                    minLineLength=params["hough"]["minLineLength"], 
+                                    maxLineGap=params["hough"]["maxLineGap"])
+    print(f"preprocess_image finished\t({time.time() - start_time:.2f} s)")
+    return img, lines
 
 # Функция для нахождения пересечения двух линий
 def find_intersection_two_lines(line1, line2):
@@ -90,29 +197,37 @@ def find_intersection_two_lines(line1, line2):
         return (int(x), int(y))
     return None
 
-def find_intersections(img, lines):
+def find_intersections(img, lines, use_debug):
+    start_time = time.time() 
+    # print("find_intersections started")
     # Находим все пересечения линий
     intersections = []
-    for i in range(len(lines)):
-        for j in range(i + 1, len(lines)):
-            line1 = lines[i][0]
-            line2 = lines[j][0]
-            intersection = find_intersection_two_lines(line1, line2)
-            if intersection:
-                intersections.append(intersection)
+    if lines is not None and len(lines) and len(lines) < 100:
+        for i in range(len(lines)):
+            for j in range(i + 1, len(lines)):
+                line1 = lines[i][0]
+                line2 = lines[j][0]
+                intersection = find_intersection_two_lines(line1, line2)
+                if intersection:
+                    intersections.append(intersection)
 
-    # Отображаем пересечения на изображении
-    output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
-    for point in intersections:
-        cv2.circle(output_img, point, 5, (0, 0, 255), -1)
+        # Отображаем результат
+        if use_debug:
+            # Отображаем пересечения на изображении
+            output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
+            for point in intersections:
+                cv2.circle(output_img, point, 5, (0, 0, 255), -1)
 
-    # Отображаем результат
-    cv2.imshow("Intersections", resize_image_to_fit(output_img))
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+            cv2.imshow("Intersections", resize_image_to_fit(output_img))
+            cv2.waitKey(0)
+            cv2.destroyAllWindows()
+    elif use_debug and lines is not None:
+        print(f"Found {len(lines)} lines!")
 
     # Преобразуем список пересечений в массив NumPy
     intersections = np.array(intersections)
+
+    print(f"find_intersections finished\t({time.time() - start_time:.2f} s)")
     return intersections
 
 
@@ -257,21 +372,32 @@ def sort_points(points):
     # Объединяем точки в нужном порядке
     return np.array([upper_points[0], upper_points[1], lower_points[0], lower_points[1]])
 
-def group_vertices3(img, intersections):
-
+def detect_squares(img, intersections, use_debug):
+    start_time = time.time()
+    # print("detect_squares started")
     # Применяем фильтрацию
     filtered_points = filter_close_points(intersections, radius=40)
+    print(f"filter_close_points completed\t({time.time() - start_time:.2f} s)")
     # Применяем усреднение
+    start_time = time.time()
     averaged_points = average_close_points(filtered_points, radius=40)
+    print(f"average_close_points completed\t({time.time() - start_time:.2f} s)")
 
-    # Отображаем усреднённые точки на изображении
-    output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
-    for point in averaged_points:
-        cv2.circle(output_img, tuple(point), 5, (0, 0, 255), -1)
+    if len(averaged_points) > 20:
+        return 20 * [0]
+    elif len(averaged_points) == 0:
+        return []
+
 
     # Отображаем результат
-    cv2.imshow("Filtered and Averaged Intersections", resize_image_to_fit(output_img))
-    cv2.waitKey(0)
+    if use_debug:
+        # Отображаем усреднённые точки на изображении
+        output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
+        for point in averaged_points:
+            cv2.circle(output_img, tuple(point), 5, (0, 0, 255), -1)
+
+        cv2.imshow("Filtered and Averaged Intersections", resize_image_to_fit(output_img))
+        cv2.waitKey(0)
 
     # Группируем вершины в квадраты
     squares = []
@@ -284,33 +410,128 @@ def group_vertices3(img, intersections):
 
     filtered_squares = filter_overlapping_squares(squares)
 
-    print(f"Найдено {len(filtered_squares)} квадратов сетки!")
-    # Отображаем квадраты на исходном изображении
-    output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
-    for square in filtered_squares:
-        color = (random.randint(0, 255), random.randint(0, 255), random.randint(0, 255))  # Случайный цвет
-        square = square.reshape((-1, 1, 2))
-        cv2.polylines(output_img, [square], isClosed=True, color=color, thickness=2)
-
     # Отображаем результат
-    cv2.imshow("Detected Squares", resize_image_to_fit(output_img))
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+    if use_debug:
+        print(f"Найдено {len(filtered_squares)} квадратов сетки!")
+        # Отображаем квадраты на исходном изображении
+        output_img = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
+        for square in filtered_squares:
+            color = (random.randint(0, 255), random.randint(0, 255), random.randint(0, 255))  # Случайный цвет
+            square = square.reshape((-1, 1, 2))
+            cv2.polylines(output_img, [square], isClosed=True, color=color, thickness=2)
 
+        cv2.imshow("Detected Squares", resize_image_to_fit(output_img))
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
+
+    # print("detect_squares finished")
     return filtered_squares
 
-def extract_grid(image_path):
-    img, lines = preprocess_image(image_path)
+def extract_grid(image_path, params):
+    use_debug = params["settings"]["debug"]
+    img, lines = preprocess_image2(image_path, params["preprocess"], use_debug)
 
-    intersections = find_intersections(img, lines)
+    intersections = find_intersections(img, lines, use_debug)
     
-    squares = group_vertices3(img, intersections)
+    squares = detect_squares(img, intersections, use_debug)
 
     return squares
 
+def optimize_params(trial, path):
+    params = {
+        "settings": {
+            "debug": False
+        },
+        "preprocess": {
+            "clahe": {
+                "to_use": trial.suggest_categorical("clahe_to_use", [True, False]),
+                "clipLimit": trial.suggest_float("clipLimit", 1.0, 8.0), 
+                "tileGridSize": trial.suggest_int("tileGridSize", 1, 5),
+            },
+            "close1": {
+                "to_use": trial.suggest_categorical("close_to_use1", [True, False]),
+                "kernel": trial.suggest_int("close_kernel1", 1, 5, step=2),
+                "iterations": trial.suggest_int("close_iterations1", 1, 5),
+            },
+            "gauss": {
+                "to_use": trial.suggest_categorical("gauss_to_use", [True, False]),
+                "kernel": trial.suggest_int("gauss_kernel", 1, 5, step=2),
+                "sigma": 0
+            },
+            "canny": {
+                "to_use": trial.suggest_categorical("canny_to_use", [True, False]),
+                "left": trial.suggest_int("left", 10, 150),
+                "right": trial.suggest_int("right", 40, 300),
+                "apertureSize": 3
+            },
+            "erode1": {
+                "to_use": trial.suggest_categorical("erode_to_use1", [True, False]),
+                "kernel": trial.suggest_int("erode_kernel1", 1, 5, step=2),
+                "iterations": trial.suggest_int("erode_iterations1", 1, 5),
+            },
+            "dilate": {
+                "to_use": trial.suggest_categorical("dilate_to_use", [True, False]),
+                "kernel": trial.suggest_int("dilate_kernel", 1, 5, step=2),
+                "iterations": trial.suggest_int("dilate_iterations", 1, 5),
+            },
+            "close2": {
+                "to_use": trial.suggest_categorical("close_to_use2", [True, False]),
+                "kernel": trial.suggest_int("close_kernel2", 1, 5, step=2),
+                "iterations": trial.suggest_int("close_iterations2", 1, 5),
+            },
+            "erode2": {
+                "to_use": trial.suggest_categorical("erode_to_use2", [True, False]),
+                "kernel": trial.suggest_int("erode_kernel2", 1, 5, step=2),
+                "iterations": trial.suggest_int("erode_iterations2", 1, 5),
+            },
+            "hough": {
+                "threshold": trial.suggest_int("threshold", 300, 800),
+                "minLineLength": trial.suggest_int("minLineLength", 100, 500), 
+                "maxLineGap": trial.suggest_int("maxLineGap", 10, 70),
+            }
+        }
+    }
+
+    mse_values = []
+    for file in sorted(os.listdir(path)):
+        grid_count = int(file.split('.')[-2])
+        full_path = os.path.join(path, file)
+        squares = extract_grid(full_path, params)
+        count = len(squares)
+        mse = (grid_count - count) ** 2
+        mse_values.append(mse)
+    return np.mean(mse_values)
+
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Поиск квадратов на изображении.")
-    parser.add_argument("-i", type=str, help="Путь к изображению.")
+    parser.add_argument("-i", type=str, help="Путь к изображениям.")
+    parser.add_argument("-c", type=str, help="Путь к конфигу с настройками.")
     args = parser.parse_args()
-    
-    extract_grid(args.i)
+    path = args.i
+    config = args.c
+
+    # start_time = time.time()
+    # study = optuna.create_study(direction="minimize")
+    # study.optimize(lambda trial: optimize_params(trial, path), n_trials=100)
+    # print(f"Лучшие параметры: {study.best_params}")
+    # print(f"Лучшее MSE: {study.best_value}")
+    # print(f"Elapsed time:{time.time() - start_time:.2f} s")
+
+    # exit(-1)
+
+    with open(config, "r") as f:
+        params = json.load(f)
+
+    mse_values = []
+    if os.path.isdir(path):
+        for file in sorted(os.listdir(path)):
+            grid_count = int(file.split('.')[-2])
+            full_path = os.path.join(path, file)
+            squares = extract_grid(full_path, params)
+            count = len(squares)
+            mse = (grid_count - count) ** 2
+            mse_values.append(mse)
+            print(f"({file})Нужно: {grid_count}, получилось: {count}")
+        print(f"MSE:", np.mean(mse_values))
+    else:
+        squares = extract_grid(args.i, params)
